added array type in connectors
