+++ duplicatesuricate/linker.py
@@ -1,19 +1,20 @@
 from connectors import _Connector
 from comparators import _Comparator
-from evaluators import _Evaluator
+from classifiers import _Classifier
+from array import _Array, _Col
 
 class _RecordLinker:
-    def __init__(self, connector, comparator, evaluator):
+    def __init__(self, connector, comparator, classifier):
         '''
 
         Args:
             connector (connectors._Connector):
             comparator (comparators._Comparator):
-            evaluator (evaluators._Evaluator):
+            classifier (classifiers._Classifier):
         '''
         self.connector = connector
         self.comparator = comparator
-        self.evaluator = evaluator
+        self.classifier = classifier
         self._coherency()
         print('init ok')
         pass
@@ -25,16 +26,10 @@ class _RecordLinker:
             bool:
         '''
         assert self.comparator.compared.issubset(self.connector.attributes)
-        assert self.evaluator.used.issubset(self.connector.relevance.union(self.comparator.scored))
+        assert self.classifier.used.issubset(self.connector.relevance.union(self.comparator.scored))
         return True
 
-    def search(self, query):
-        return self.connector.search(query)
-
-    def compare(self,query, targets):
-        return self.comparator.compare(query, targets)
-
-    def predict_proba(self, query, on_index=None, return_filtered=True):
+    def predict_proba(self, query, on_index=None):
         """
         Main method of this class:
         - with the help of the scoring model, create a similarity table
@@ -42,36 +37,36 @@ class _RecordLinker:
         - returns that probability
 
         Args:
-            query (pd.Series): information available on our query
+            query: information available on our query
             on_index (pd.Index): index on which to do the prediction
-            return_filtered (bool): whether or not to filter the table
         Returns:
-            pd.Series : the probability vector of the target records being the same as the query
+            _Col: the probability vector of the target records being the same as the query
 
         """
-        # TODO
-        output = self.connector.search(query)
+        output = self.connector.search(query, on_index=on_index)
 
-        relevance = output.select(self.connector.relevance)
-        targets = output.select(self.connector.attributes)
-
-        comparison =
-
-        if table_score_complete is None or table_score_complete.shape[0] == 0:
+        if output is None or output.count() == 0:
             return None
         else:
-            # launch prediction using the predict_proba of the scikit-learn module
 
-            y_proba = self.classifier.predict_proba(table_score_complete).copy()
+            relevance = output.select(self.connector.relevance)
+            targets = output.select(self.connector.attributes)
 
-            del table_score_complete
+            # create table of scores
+            scores = self.comparator.compare(query, targets=targets)
+            scores = scores.union(relevance)
+
+            # launch prediction
+            y_proba = self.classifier.predict_proba(scores)
 
             # sort the results
-            y_proba.sort_values(ascending=False, inplace=True)
+            y_proba = y_proba.sort_values(ascending=False)
+
+            del scores
 
             return y_proba
 
 es = _Connector()
 fz = _Comparator()
-ml = _Evaluator()
-rl = _RecordLinker(connector=es, comparator=fz, evaluator=ml)
+ml = _Classifier()
+rl = _RecordLinker(connector=es, comparator=fz, classifier=ml)
